Move test input class to a package to test nested lookups
