+++ tests/input_files/mypackage/cmodule.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+from typing import Optional, Tuple
+
+
+@dataclass
+class ExampleClass:
+
+    def __init__(self, a: Optional[Tuple[int, int]] = None, name: Optional[str] = None):
+        self.a = a
+        self.name = name
